improve slot context
