+++ modules/gui-bukkit/src/main/kotlin/ru/astrainteractive/astramarket/gui/button/BorderButton.kt
@@ -1,0 +1,14 @@
+package ru.astrainteractive.astramarket.gui.button
+
+import ru.astrainteractive.astralibs.menu.slot.InventorySlot
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setDisplayName
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setIndex
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setItemStack
+import ru.astrainteractive.astramarket.gui.button.di.ButtonContext
+import ru.astrainteractive.astramarket.gui.util.ItemStackExt.toItemStack
+
+internal fun ButtonContext.border(index: Int) = InventorySlot.Builder()
+    .setIndex(index)
+    .setItemStack(config.buttons.border.toItemStack())
+    .setDisplayName(" ")
+    .build()
