+++ modules/gui-bukkit/src/main/kotlin/ru/astrainteractive/astramarket/gui/button/PrevPageButton.kt
@@ -1,0 +1,20 @@
+package ru.astrainteractive.astramarket.gui.button
+
+import ru.astrainteractive.astralibs.menu.clicker.Click
+import ru.astrainteractive.astralibs.menu.slot.InventorySlot
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.editMeta
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setIndex
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setItemStack
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setOnClickListener
+import ru.astrainteractive.astramarket.gui.button.di.ButtonContext
+import ru.astrainteractive.astramarket.gui.util.ItemStackExt.toItemStack
+
+internal fun ButtonContext.prevPage(
+    index: Int,
+    click: Click
+) = InventorySlot.Builder()
+    .setIndex(index)
+    .setItemStack(config.buttons.previous.toItemStack())
+    .editMeta { displayName(translation.menu.prev.component) }
+    .setOnClickListener(click)
+    .build()
