+++ modules/gui-bukkit/src/main/kotlin/ru/astrainteractive/astramarket/gui/players/PlayersGui.kt
@@ -18,7 +18,13 @@ import ru.astrainteractive.astralibs.menu.inventory.util.PaginatedInventoryMenuE
 import ru.astrainteractive.astralibs.menu.inventory.util.PaginatedInventoryMenuExt.showPage
 import ru.astrainteractive.astralibs.menu.inventory.util.PaginatedInventoryMenuExt.showPrevPage
 import ru.astrainteractive.astralibs.menu.slot.InventorySlot
-import ru.astrainteractive.astramarket.gui.button.di.MenuDrawerContext
+import ru.astrainteractive.astramarket.gui.button.back
+import ru.astrainteractive.astramarket.gui.button.di.ButtonContext
+import ru.astrainteractive.astramarket.gui.button.expired
+import ru.astrainteractive.astramarket.gui.button.nextPage
+import ru.astrainteractive.astramarket.gui.button.playerItem
+import ru.astrainteractive.astramarket.gui.button.playersSort
+import ru.astrainteractive.astramarket.gui.button.prevPage
 import ru.astrainteractive.astramarket.gui.di.AuctionGuiDependencies
 import ru.astrainteractive.astramarket.gui.invmap.AuctionInventoryMap
 import ru.astrainteractive.astramarket.gui.invmap.AuctionInventoryMap.AuctionSlotKey
@@ -34,11 +40,10 @@ internal class PlayersGui(
     private val playersMarketComponent: PlayersMarketComponent,
     player: Player,
     dependencies: AuctionGuiDependencies,
-    menuDrawerContext: MenuDrawerContext
+    private val buttonContext: ButtonContext
 ) : PaginatedInventoryMenu(),
     AuctionGuiDependencies by dependencies,
-    KyoriComponentSerializer by dependencies.kyoriComponentSerializer,
-    MenuDrawerContext by menuDrawerContext {
+    KyoriComponentSerializer by dependencies.kyoriComponentSerializer {
     override val inventorySize: InventorySize = InventorySize.XL
 
     private val inventoryMap: AuctionInventoryMap
@@ -55,7 +60,7 @@ internal class PlayersGui(
     )
 
     override val prevPageButton: InventorySlot
-        get() = prevPageButtonFactory.render(
+        get() = buttonContext.prevPage(
             index = inventoryMap.indexOf(AuctionSlotKey.PR),
             click = {
                 playerHolder.player.playSound(config.sounds.open)
@@ -64,7 +69,7 @@ internal class PlayersGui(
         )
 
     override val nextPageButton: InventorySlot
-        get() = nextPageButtonFactory.render(
+        get() = buttonContext.nextPage(
             index = inventoryMap.indexOf(AuctionSlotKey.NE),
             click = {
                 playerHolder.player.playSound(config.sounds.open)
@@ -73,7 +78,7 @@ internal class PlayersGui(
         )
 
     private val sortButton: InventorySlot
-        get() = playersSortButtonFactory.render(
+        get() = buttonContext.playersSort(
             index = inventoryMap.indexOf(AuctionSlotKey.FI),
             sortType = playersMarketComponent.model.value.sort,
             click = {
@@ -86,7 +91,7 @@ internal class PlayersGui(
         )
 
     private val expiredButton: InventorySlot
-        get() = expiredButtonFactory.render(
+        get() = buttonContext.expired(
             index = inventoryMap.indexOf(AuctionSlotKey.AU),
             isExpired = playersMarketComponent.model.value.isExpired,
             click = {
@@ -96,7 +101,7 @@ internal class PlayersGui(
         )
 
     private val closeButton: InventorySlot
-        get() = backButtonFactory.render(
+        get() = buttonContext.back(
             index = inventoryMap.indexOf(AuctionSlotKey.BA),
             click = { playerHolder.player.closeInventory() }
         )
@@ -113,7 +118,7 @@ internal class PlayersGui(
                     .playersAndSlots
                     .filter { it.slots.any { slot -> slot.expired == isExpired } }
                     .getOrNull(index) ?: return@withKeySlot null
-                playerItemButtonFactory.render(
+                buttonContext.playerItem(
                     playerAndSlots = items,
                     index = slotIndex,
                     isExpired = playersMarketComponent.model.value.isExpired,
