+++ modules/gui-bukkit/src/main/kotlin/ru/astrainteractive/astramarket/gui/slots/SlotsGui.kt
@@ -24,7 +24,15 @@ import ru.astrainteractive.astralibs.menu.inventory.util.PaginatedInventoryMenuE
 import ru.astrainteractive.astralibs.menu.slot.InventorySlot
 import ru.astrainteractive.astralibs.permission.BukkitPermissibleExt.toPermissible
 import ru.astrainteractive.astramarket.core.PluginPermission
-import ru.astrainteractive.astramarket.gui.button.di.MenuDrawerContext
+import ru.astrainteractive.astramarket.gui.button.aauc
+import ru.astrainteractive.astramarket.gui.button.auctionSort
+import ru.astrainteractive.astramarket.gui.button.back
+import ru.astrainteractive.astramarket.gui.button.border
+import ru.astrainteractive.astramarket.gui.button.di.ButtonContext
+import ru.astrainteractive.astramarket.gui.button.expired
+import ru.astrainteractive.astramarket.gui.button.expiredSlot
+import ru.astrainteractive.astramarket.gui.button.nextPage
+import ru.astrainteractive.astramarket.gui.button.prevPage
 import ru.astrainteractive.astramarket.gui.di.AuctionGuiDependencies
 import ru.astrainteractive.astramarket.gui.invmap.AuctionInventoryMap
 import ru.astrainteractive.astramarket.gui.invmap.AuctionInventoryMap.AuctionSlotKey
@@ -39,12 +47,11 @@ import ru.astrainteractive.astramarket.market.presentation.AuctionComponent
 internal class SlotsGui(
     player: Player,
     dependencies: AuctionGuiDependencies,
-    menuDrawerContext: MenuDrawerContext,
+    private val buttonContext: ButtonContext,
     private val auctionComponent: AuctionComponent
 ) : PaginatedInventoryMenu(),
     AuctionGuiDependencies by dependencies,
-    KyoriComponentSerializer by dependencies.kyoriComponentSerializer,
-    MenuDrawerContext by menuDrawerContext {
+    KyoriComponentSerializer by dependencies.kyoriComponentSerializer {
     override val playerHolder = DefaultPlayerHolder(player)
     override val title: Component = let {
         val playerNameComponent = auctionComponent.model.value
@@ -69,23 +76,23 @@ internal class SlotsGui(
     private val borderButtons: List<InventorySlot>
         get() = inventoryMap.withKeySlot(
             key = AuctionSlotKey.BO,
-            transform = borderButtonRenderer::render
+            transform = buttonContext::border
         )
 
     override val nextPageButton: InventorySlot
-        get() = nextPageButtonFactory.render(
+        get() = buttonContext.nextPage(
             index = inventoryMap.indexOf(AuctionSlotKey.NE),
             click = { onNextPageClicked() }
         )
 
     override val prevPageButton: InventorySlot
-        get() = prevPageButtonFactory.render(
+        get() = buttonContext.prevPage(
             index = inventoryMap.indexOf(AuctionSlotKey.PR),
             click = { onPrevPageClicked() }
         )
 
     private val sortButton: InventorySlot
-        get() = auctionSortButtonFactory.render(
+        get() = buttonContext.auctionSort(
             index = inventoryMap.indexOf(AuctionSlotKey.FI),
             sortType = auctionComponent.model.value.sortType,
             click = {
@@ -95,7 +102,7 @@ internal class SlotsGui(
         )
 
     private val expiredButton: InventorySlot
-        get() = expiredButtonFactory.render(
+        get() = buttonContext.expired(
             index = inventoryMap.indexOf(AuctionSlotKey.AU),
             isExpired = true,
             click = {
@@ -105,13 +112,13 @@ internal class SlotsGui(
         )
 
     private val aaucButton: InventorySlot
-        get() = aaucButtonFactory.render(
+        get() = buttonContext.aauc(
             index = inventoryMap.indexOf(AuctionSlotKey.AU),
             click = { auctionComponent.toggleExpired() }
         )
 
     private val openPlayersButton: InventorySlot
-        get() = backButtonFactory.render(
+        get() = buttonContext.back(
             index = inventoryMap.indexOf(AuctionSlotKey.BA),
             click = {
                 val route = GuiRouter.Route.Players(
@@ -123,7 +130,7 @@ internal class SlotsGui(
         )
 
     private val closeButton: InventorySlot
-        get() = backButtonFactory.render(
+        get() = buttonContext.back(
             index = inventoryMap.indexOf(AuctionSlotKey.BA),
             click = { playerHolder.player.closeInventory() }
         )
@@ -156,7 +163,7 @@ internal class SlotsGui(
                     .getOrNull(index)
                     ?: return@withKeySlot null
                 val permissible = playerHolder.player.toPermissible()
-                expiredMarketItemButtonFactory.render(
+                buttonContext.expiredSlot(
                     auctionItem = auctionItem,
                     index = slotIndex,
                     click = { onAuctionItemClicked(index, it.click) },
